First backmerge from trunk to 0-9 branch for Java. Not all java tests passing yet
git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/branches/qpid.0-9@507672 13f79535-47bb-0310-9956-ffa450edef68

+++ java/common/src/main/java/org/apache/qpid/framing/RequestManager.java
@@ -40,7 +40,7 @@ public class RequestManager
      * to be known.
      */
     private boolean serverFlag;
-    private int connectionId;
+    private long connectionId;
 
     /**
      * Request and response frames must have a requestID and responseID which
@@ -56,7 +56,7 @@ public class RequestManager
 
     private ConcurrentHashMap<Long, AMQMethodListener> requestSentMap;
 
-    public RequestManager(int connectionId, int channel, AMQProtocolWriter protocolWriter, boolean serverFlag)
+    public RequestManager(long connectionId, int channel, AMQProtocolWriter protocolWriter, boolean serverFlag)
     {
         this.channel = channel;
         this.protocolWriter = protocolWriter;
